// src/docs/pbes/node/point_pb.ts
// @generated by protoc-gen-es v2.10.2 with parameter "target=ts,js_import_style=legacy_commonjs"
// @generated from file node/point.proto (package yuhaiin.point, edition 2023)


import type { GenEnum, GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { enumDesc, fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { protocol } from "./protocol_pb";
import { file_node_protocol } from "./protocol_pb";
import { file_google_protobuf_go_features } from "@bufbuild/protobuf/wkt";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file node/point.proto.
 */
export const file_node_point: GenFile = /*@__PURE__*/
  fileDesc("ChBub2RlL3BvaW50LnByb3RvEg15dWhhaWluLnBvaW50IowBCgVwb2ludBIMCgRoYXNoGAEgASgJEgwKBG5hbWUYAiABKAkSDQoFZ3JvdXAYAyABKAkSJQoGb3JpZ2luGAQgASgOMhUueXVoYWlpbi5wb2ludC5vcmlnaW4SMQoJcHJvdG9jb2xzGAUgAygLMhoueXVoYWlpbi5wcm90b2NvbC5wcm90b2NvbEICKAEqLQoGb3JpZ2luEgsKB3Jlc2VydmUQABIKCgZyZW1vdGUQZRIKCgZtYW51YWwQZkI2WixnaXRodWIuY29tL0FzdXRvcnVmYS95dWhhaWluL3BrZy9wcm90b3Mvbm9kZZIDBdI+AhADYghlZGl0aW9uc3DoBw", [file_node_protocol, file_google_protobuf_go_features]);

/**
 * @generated from message yuhaiin.point.point
 */
export type point = Message<"yuhaiin.point.point"> & {
  /**
   * @generated from field: string hash = 1;
   */
  hash: string;

  /**
   * @generated from field: string name = 2;
   */
  name: string;

  /**
   * @generated from field: string group = 3;
   */
  group: string;

  /**
   * @generated from field: yuhaiin.point.origin origin = 4;
   */
  origin: origin;

  /**
   * will use protocols' order to create dialer
   *
   * @generated from field: repeated yuhaiin.protocol.protocol protocols = 5;
   */
  protocols: protocol[];
};

/**
 * Describes the message yuhaiin.point.point.
 * Use `create(pointSchema)` to create a new message.
 */
export const pointSchema: GenMessage<point> = /*@__PURE__*/
  messageDesc(file_node_point, 0);

/**
 * @generated from enum yuhaiin.point.origin
 */
export enum origin {
  /**
   * @generated from enum value: reserve = 0;
   */
  reserve = 0,

  /**
   * @generated from enum value: remote = 101;
   */
  remote = 101,

  /**
   * @generated from enum value: manual = 102;
   */
  manual = 102,
}

/**
 * Describes the enum yuhaiin.point.origin.
 */
export const originSchema: GenEnum<origin> = /*@__PURE__*/
  enumDesc(file_node_point, 0);